# swaggerDocs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "File-Sharing-APIs",
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/files": {
      "post": {
        "summary": "Upload Files",
        "tags": [
          "Misc"
        ],
        "operationId": "UploadFiles",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "myFile",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/files/9185c0ea-e07e-4d11-90dd-fc9ef760f1b8": {
      "get": {
        "summary": "Download Page",
        "tags": [
          "Misc"
        ],
        "operationId": "DownloadPage",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/files/download/a63236a8-fadc-460c-9f36-6d467cfdb865": {
      "get": {
        "summary": "Download file",
        "tags": [
          "Misc"
        ],
        "operationId": "Downloadfile",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/files/send": {
      "post": {
        "summary": "Send Email",
        "tags": [
          "Misc"
        ],
        "operationId": "SendEmail",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SendEmailRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    }
  },
  "definitions": {
    "SendEmailRequest": {
      "title": "SendEmailRequest",
      "example": {
        "uuid": "08487044-a3f0-4c73-8017-04e9cd50a052",
        "emailTo": "hkanani191@gmail.com",
        "emailFrom": "harikanani2003@gmail.com"
      },
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "emailTo": {
          "type": "string"
        },
        "emailFrom": {
          "type": "string"
        }
      },
      "required": [
        "uuid",
        "emailTo",
        "emailFrom"
      ]
    }
  },
  "tags": [
    {
      "name": "Misc",
      "description": ""
    }
  ]
}